requirements.txt
easyocr==1.7.1
streamlit==1.32.2
pillow==10.2.0